requirements.txt
anthropic>=0.19.1
tqdm>=4.66.1
python-dotenv>=1.0.0
